requirements.txt
Django==1.10.3
django-environ==0.4.1
djangorestframework==3.5.3
psycopg2==2.6.2
requests==2.12.1
six==1.10.0
